fastapi/requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
h11==0.14.0
idna==3.4
imageio==2.25.0
install==1.3.5
numpy==1.24.2
opencv-python==4.7.0.68
Pillow==9.4.0
pydantic==1.10.4
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
websockets==10.4
